Modify the cluster health check task to filter out any V4 jobs
The V4 jobs will be running via the Agent and won't have a healthcheck endpoint to call. This is a v3 remnant and once all jobs are moved to v4 logic the ClusterHealthCheck may be re-written to factor in other factors. For now it needs to stay the way it behaves to maintain v3 job api functionality.